+++ index.html
@@ -21,11 +21,13 @@
         <div id="end-message"></div>
         <button id="restart-button">Restart Game</button>
 
-        <script src="game.js"></script>
-        <script src="levels.js"></script>
-        <script>
-            runGame(GAME_LEVELS, DOMDisplay);
-        </script>
+        <script type="module" src="dom.js"></script>
+        <script type="module" src="core.js"></script>
+        
+        <script type="module" src="gameLogic.js"></script>
+        <script type="module" src="utils.js"></script>
+        <script type="module" src="levels.js"></script>
+        <script type="module" src="main.js"></script>
   </body>
 
 </html>
