+++ levels.js
@@ -1,0 +1,31 @@
+export const GAME_LEVELS = [`                                                    
+................................................................................
+................................................................................
+................................................................................
+................................................................................
+................................................................................
+................................................................................
+..................................................................###...........
+...................................................##......##....##+##..........
+....................................o.o......##..................#+++#..........
+.................................................................##+##..........
+...................................#####..........................#v#...........
+............................................................................##..
+..##......................................o.o................................#..
+..#.....................o....................................................#..
+..#......................................#####.............................o.#..
+..#..........####.......o....................................................#..
+..#..@.......#..#................................................#####.......#..
+..############..###############...####################.....#######...#########..
+..............................#...#..................#.....#....................
+..............................#+++#..................#+++++#....................
+..............................#+++#..................#+++++#....................
+..............................#####..................#######....................
+................................................................................
+................................................................................
+`];
+
+// if (typeof module != "undefined" && module.exports && (typeof window == "undefined" || window.exports != exports))
+//   module.exports = GAME_LEVELS;
+// if (typeof global != "undefined" && !global.GAME_LEVELS)
+//   global.GAME_LEVELS = GAME_LEVELS;
