--- src/levels.js
@@ -1,178 +1,0 @@
-var GAME_LEVELS = [`                                                    
-................................................................................
-................................................................................
-................................................................................
-................................................................................
-................................................................................
-................................................................................
-..................................................................###...........
-...................................................##......##....##+##..........
-....................................o.o......##..................#+++#..........
-.................................................................##+##..........
-...................................#####..........................#v#...........
-............................................................................##..
-..##......................................o.o................................#..
-..#.....................o....................................................#..
-..#......................................#####.............................o.#..
-..#..........####.......o....................................................#..
-..#..@.......#..#................................................#####.......#..
-..############..###############...####################.....#######...#########..
-..............................#...#..................#.....#....................
-..............................#+++#..................#+++++#....................
-..............................#+++#..................#+++++#....................
-..............................#####..................#######....................
-................................................................................
-................................................................................
-`,`                                                                     
-................................................................................
-................................................................................
-....###############################.............................................
-...##.............................##########################################....
-...#.......................................................................##...
-...#....o...................................................................#...
-...#................................................=.......................#...
-...#.o........################...................o..o...........|........o..#...
-...#.........................#..............................................#...
-...#....o....................##########.....###################....##########...
-...#..................................#+++++#.................#....#............
-...###############....oo......=o.o.o..#######.###############.#....#............
-.....#...............o..o.............#.......#......#........#....#............
-.....#....................#############..######.####.#.########....########.....
-.....#.............########..............#...........#.#..................#.....
-.....#..........####......####...#####################.#..................#.....
-.....#........###............###.......................########....########.....
-.....#.......##................#########################......#....#............
-.....#.......#................................................#....#............
-.....###......................................................#....#............
-.......#...............o...........................................#............
-.......#...............................................o...........#............
-.......#########......###.....############.........................##...........
-.............#..................#........#####....#######.o.........########....
-.............#++++++++++++++++++#............#....#.....#..................#....
-.............#++++++++++++++++++#..........###....###...####.o.............#....
-.............####################..........#........#......#.....|.........#....
-...........................................#++++++++#......####............#....
-...........................................#++++++++#.........#........@...#....
-...........................................#++++++++#.........##############....
-...........................................##########...........................
-................................................................................
-`,`
-......................................#++#........................#######....................................#+#..
-......................................#++#.....................####.....####.................................#+#..
-......................................#++##########...........##...........##................................#+#..
-......................................##++++++++++##.........##.............##...............................#+#..
-.......................................##########++#.........#....................................o...o...o..#+#..
-................................................##+#.........#.....o...o....................................##+#..
-.................................................#+#.........#................................###############++#..
-.................................................#v#.........#.....#...#........................++++++++++++++##..
-.............................................................##..|...|...|..##............#####################...
-..............................................................##+++++++++++##............v........................
-...............................................................####+++++####......................................
-...............................................#.....#............#######........###.........###..................
-...............................................#.....#...........................#.#.........#.#..................
-...............................................#.....#.............................#.........#....................
-...............................................#.....#.............................##........#....................
-...............................................##....#.............................#.........#....................
-...............................................#.....#......o..o.....#...#.........#.........#....................
-...............#######........###...###........#.....#...............#...#.........#.........#....................
-..............##.....##.........#...#..........#.....#.....######....#...#...#########.......#....................
-.............##.......##........#.o.#..........#....##...............#...#...#...............#....................
-.....@.......#.........#........#...#..........#.....#...............#...#...#...............#....................
-....###......#.........#........#...#..........#.....#...............#...#####...######......#....................
-....#.#......#.........#.......##.o.##.........#.....#...............#.....o.....#.#.........#....................
-++++#.#++++++#.........#++++++##.....##++++++++##....#++++++++++.....#.....=.....#.#.........#....................
-++++#.#++++++#.........#+++++##.......##########.....#+++++++##+.....#############.##..o.o..##....................
-++++#.#++++++#.........#+++++#....o.................##++++++##.+....................##.....##.....................
-++++#.#++++++#.........#+++++#.....................##++++++##..+.....................#######......................
-++++#.#++++++#.........#+++++##.......##############++++++##...+..................................................
-++++#.#++++++#.........#++++++#########++++++++++++++++++##....+..................................................
-++++#.#++++++#.........#++++++++++++++++++++++++++++++++##.....+..................................................
-`,`
-..............................................................................................................
-..............................................................................................................
-..............................................................................................................
-..............................................................................................................
-..............................................................................................................
-........................................o.....................................................................
-..............................................................................................................
-........................................#.....................................................................
-........................................#.....................................................................
-........................................#.....................................................................
-........................................#.....................................................................
-.......................................###....................................................................
-.......................................#.#.................+++........+++..###................................
-.......................................#.#.................+#+........+#+.....................................
-.....................................###.###................#..........#......................................
-......................................#...#.................#...oooo...#.......###............................
-......................................#...#.................#..........#......#+++#...........................
-......................................#...#.................############.......###............................
-.....................................##...##......#...#......#................................................
-......................................#...#########...########..............#.#...............................
-......................................#...#...........#....................#+++#..............................
-......................................#...#...........#.....................###...............................
-.....................................##...##..........#.......................................................
-......................................#...#=.=.=.=....#............###........................................
-......................................#...#...........#...........#+++#.......................................
-......................................#...#....=.=.=.=#.....o......###.......###..............................
-.....................................##...##..........#.....................#+++#.............................
-..............................o...o...#...#...........#.....#................##v........###...................
-......................................#...#...........#..............#.................#+++#..................
-.............................###.###.###.###.....o.o..#++++++++++++++#...................v#...................
-.............................#.###.#.#.###.#..........#++++++++++++++#........................................
-.............................#.............#...#######################........................................
-.............................##...........##.........................................###......................
-..###.........................#.....#.....#.........................................#+++#................###..
-..#.#.........................#....###....#..........................................###.................#.#..
-..#...........................#....###....#######........................#####.............................#..
-..#...........................#...........#..............................#...#.............................#..
-..#...........................##..........#..............................#.#.#.............................#..
-..#.......................................#.......|####|....|####|.....###.###.............................#..
-..#................###.............o.o....#..............................#.........###.....................#..
-..#...............#####.......##..........#.............................###.......#+++#..........#.........#..
-..#...............o###o.......#....###....#.............................#.#........###..........###........#..
-..#................###........#############..#.oo.#....#.oo.#....#.oo..##.##....................###........#..
-..#......@..........#.........#...........#++#....#++++#....#++++#....##...##....................#.........#..
-..#############################...........#############################.....################################..
-..............................................................................................................
-..............................................................................................................
-`,`
-..................................................................................................###.#.......
-......................................................................................................#.......
-..................................................................................................#####.......
-..................................................................................................#...........
-..................................................................................................#.###.......
-..........................o.......................................................................#.#.#.......
-.............................................................................................o.o.o###.#.......
-...................###................................................................................#.......
-.......+..o..+................................................#####.#####.#####.#####.#####.#####.#####.......
-.......#.....#................................................#...#.#...#.#...#.#...#.#...#.#...#.#...........
-.......#=.o..#............#...................................###.#.###.#.###.#.###.#.###.#.###.#.#####.......
-.......#.....#..................................................#.#...#.#...#.#...#.#...#.#...#.#.....#.......
-.......+..o..+............o..................................####.#####.#####.#####.#####.#####.#######.......
-..............................................................................................................
-..........o..............###..............................##..................................................
-..............................................................................................................
-..............................................................................................................
-......................................................##......................................................
-...................###.........###............................................................................
-..............................................................................................................
-..........................o.....................................................#......#......................
-..........................................................##.....##...........................................
-.............###.........###.........###.................................#..................#.................
-..............................................................................................................
-.................................................................||...........................................
-..###########.................................................................................................
-..#.........#.o.#########.o.#########.o.##................................................#...................
-..#.........#...#.......#...#.......#...#.................||..................#.....#.........................
-..#..@......#####...o...#####...o...#####.....................................................................
-..#######.....................................#####.......##.....##.....###...................................
-........#=..................=................=#...#.....................###...................................
-........#######################################...#+++++++++++++++++++++###+++++++++++++++++++++++++++++++++++
-..................................................############################################################
-..............................................................................................................
-`];
-
-if (typeof module != "undefined" && module.exports && (typeof window == "undefined" || window.exports != exports))
-  module.exports = GAME_LEVELS;
-if (typeof global != "undefined" && !global.GAME_LEVELS)
-  global.GAME_LEVELS = GAME_LEVELS;
