+++ utils.js
@@ -1,0 +1,37 @@
+import { Player } from "./core.js";
+import { Coin } from "./core.js";
+import { Lava } from "./core.js";
+
+function parseQuery(queryString) {
+    var query = {};
+    var pairs = (queryString[0] === '?' ? queryString.substr(1) : queryString).split('&');
+    for (var i = 0; i < pairs.length; i++) {
+      var pair = pairs[i].split('=');
+      query[decodeURIComponent(pair[0])] = decodeURIComponent(pair[1] || '');
+    }
+    return query;
+}
+
+function trackKeys(keys) {
+    let down = Object.create(null);
+    function track(event) {
+      if (keys.includes(event.key)) {
+        down[event.key] = event.type == "keydown";
+        event.preventDefault();
+      }
+    }
+    window.addEventListener("keydown", track);
+    window.addEventListener("keyup", track);
+    return down;
+  }
+  
+var arrowKeys = trackKeys(["ArrowLeft", "ArrowRight", "ArrowUp"]);
+
+const levelChars = {
+    ".": "empty", "#": "wall", "+": "lava",
+    "@": Player, "o": Coin,
+    "=": Lava, "|": Lava, "v": Lava
+};
+
+
+export { parseQuery, arrowKeys, levelChars };
