+++ vite.config.js
@@ -1,0 +1,7 @@
+// vite.config.js
+export default {
+  root: './',
+  build: {
+    outDir: 'dist'
+  }
+}
